+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/2)
 * Implement styled-components instead .css files
+* Add tooltips for navigation menu items
 
 0.11 (2024-05-02)
 ===============